backend/requirements.txt
# ─── Web framework ────────────────────────────────────────
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# ─── Database ─────────────────────────────────────────────
sqlalchemy==2.0.35
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.3
pgvector==0.3.3

# ─── Auth0 ────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
httpx==0.27.2

# ─── LLM & Parsing ────────────────────────────────────────
google-generativeai==0.8.3
llama-parse==0.5.14
llama-index==0.11.20
llama-index-vector-stores-postgres==0.2.5
llama-index-embeddings-gemini==0.2.1
llama-index-llms-gemini==0.3.7

# ─── Scraping ─────────────────────────────────────────────
playwright==1.47.0
beautifulsoup4==4.12.3
requests==2.32.3

# ─── Utilities ────────────────────────────────────────────
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
difflib2==1.0.0
hashlib2==1.0.0
apscheduler==3.10.4
